Meal Price
This is a little project from that I worked on collage on my first week learning Python. It calculates the total amount to pay for food in a restaurant. It asks for User input and outputs the Total amount to pay according to your answer. Also, I added a feature to ask if the person wants to pay the Brazilian Service Tax which is generally 10% of the total amount to pay.